+++ src/Interfaces/IExchangeService.cs
@@ -3,5 +3,5 @@ using CurrencyExchange.Api.Models.Responses;
 namespace CurrencyExchange.Api.Interfaces;
 
 public interface IExchangeService {
-    ExchangeResultResponse Exchange(string baseCurrencyCode, string targetCurrencyCode, double amount);
+    ExchangeResultResponse Exchange(string baseCurrencyCode, string targetCurrencyCode, decimal amount);
 }
